api/requirements.txt
fastapi
uvicorn[standard]
influxdb-client
python-dotenv
